Refactor UiLayout component to enhance maintainability and user experience
- Removed unused imports and state management related to theme and pathname.
- Consolidated header elements into a new Navbar component for improved structure and readability.
- Streamlined the layout by directly integrating essential components, enhancing overall code organization.
